backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
pandas
openpyxl
python-dotenv
python-multipart
rapidfuzz
python-dateutil
pytest
